+++ docs/media/webview.js
@@ -12,7 +12,8 @@
   mermaidTheme,
   vegaTheme,
   plantumlRenderMode,
-  plantumlWebsite
+  plantumlWebsite,
+  katexDelimiters
 ) {
   class ContextMenu {
     constructor(selector, menuItems) {
@@ -147,6 +148,7 @@
         options: {
           markdown: { flavor: markdownFlavor },
           vega: { theme: vegaTheme },
+          katex: { delimiters: katexDelimiters },
           mermaid: { theme: mermaidTheme },
           plantuml: { renderMode: plantumlRenderMode, umlWebSite: plantumlWebsite }
         }
@@ -165,10 +167,27 @@
 
     updateOptions() {
       if (this.config.options.markdown.flavor) {
-        previewer.addOptions(this.config.options.markdown);
+        previewer.setShowdownFlavor(this.config.options.markdown.flavor);
       }
       if (this.config.options.vega.theme) {
         previewer.setVegaOptions(Object.assign(this.config.options.vega, { renderer: 'svg' }));
+      }
+      if (this.config.options.katex.delimiters) {
+        const options = JSON.parse('[' + this.config.options.katex.delimiters + ']');
+        if (Array.isArray(options)) {
+          let delimiters = [];
+          options.forEach((option) => {
+            if (option.hasOwnProperty('left') && option.hasOwnProperty('right')) {
+              if (!option.hasOwnProperty('display')) {
+                option.display = false;
+              }
+              delimiters.push(option);
+            }
+          });
+          if (delimiters.length) {
+            previewer.setKatexOptions({ delimiters: delimiters });
+          }
+        }
       }
       if (this.config.options.mermaid.theme) {
         previewer.setMermaidOptions(this.config.options.mermaid);
@@ -484,5 +503,6 @@
   typeof mermaid_theme === 'undefined' ? 'default' : mermaid_theme,
   typeof vega_theme === 'undefined' ? 'vox' : vega_theme,
   typeof plantuml_rendermode === 'undefined' ? 'local' : plantuml_rendermode,
-  typeof plantuml_website === 'undefined' ? 'www.plantuml.com/plantuml' : plantuml_website
+  typeof plantuml_website === 'undefined' ? 'www.plantuml.com/plantuml' : plantuml_website,
+  typeof katex_delimiters === 'undefined' ? '' : katex_delimiters
 );